[Many Features] - V2 (#81)
* Refactor

* Fix tests

* Split up apply payment

* Fix

* Rename prepay to collectPayment

* Rename prepayment to collectedPayment everywhere

* Fix eslint preventing client commits

* Fix client references to prepayment

* Better way to ignore the client directory

* Fix late payments logic

* [Feature] - Access Control (#67)

* Got all tests passing after doing the ACL upgrade.

- There are not real semantic changes here. Just an implementation
difference of using AccessControl instead of Ownable.

- This has numerous advantages, including the fact that we no longer
have to create methods on the CreditDesk if we want to change any admin
abilities of the Pool itself. Because now we can have multiple owners
of the Pool.

- This moves all of our contracts to an "AccessControl" model,
which allows for multiple addresses to hold a given role, and gives
us the flexibility to create new roles over time.

- Also moved a lot of repeated logic into a "BaseContract", which gives us
upgradeability, access control, and pausability all in one little package.

* Take our first win from the upgrade.

- Get rid of the setPoolTotalFundsLimit on CreditDesk.
- We only had that because we couldn't directly call the pool for admin functions, except
through the CreditDesk. Now that we can set multiple owners, we can!

* Remove the .only!

* Run CI stuff on all branches.

* I think this is the right syntax.

* hopefully this is the right syntax

* Now run lint in the github jobs.

* Fix a linter error.

* Only do the push changes for master branch.

* Fix a few small linting things.


* Upgradable credit line via minimal proxy

* Fixes

* Tests

* Use contract alias instead of address

* Minor cleanup

* [Feature] - Config contract (#68)

* Take our first win from the upgrade.

- Get rid of the setPoolTotalFundsLimit on CreditDesk.
- We only had that because we couldn't directly call the pool for admin functions, except
through the CreditDesk. Now that we can set multiple owners, we can!

* Getting the config branch started

* [WIP] - Adding the config contract.

* Got more of the Config Contract branch working...

- Added more tests.
- Added it into the deployment scripts.
- Added TotalFundsLimit as well.

* Got all tests passing for part of the Config Contract refactor.

* Fix linting things.

* Added the first address use case of the Config Contract.

- Also added a new test file just for the config contract.

* Fixed up the deployment for the CreditDesk to work with the Config Contract.

* Fix up linter issues.

* Fixed the front-ends so they work now.

* Actually make the front-ends work.

- Also remove a stray .only

* Add an additional test to ensure the enum ordering doesn't change.

* Fix a few linting things.


* Additional config refactors

* Pull even more config values into the contract

* Fixes

* Fix authorization issue on pool

* Ensure addresses can only be set once unless forced

* Fixes

* [Feature] - Add Fidu, and other refactors (#66)

* Got initial Fidu tests passing.

- This also refactors our deployment and testing system to use the same functions.
- That refactor has several benefits. One, it ensures we use the same deployment in
tests as we do in prod (minor testing details not withstanding). It also greatly improves
   test speed (from 15s to 5s for the Pool tests).
- It adds the Fidu token, which is based on an OpenZeppelin preset of an ERC20 with access controls and pausability.
- It adds tests.
- It removes continuing upkeep of our total shares struct, and that now moves to Fidu.
- This will require a migration to Fidu, which I'll write later.

* Fix up the branch from the merge so tests are passing again.

- Also improved our testing to always re-use our base deploy.
- Also added fixtures for our test setup, which greatly improves test speed.

* Converting Fidu to use the Config contract approach.

* Fixing some front-end things to prepare for using Fidu

* Got all tests passing on the Fidu branch!

- Also cleaned up a lot of the deploy scripts
- And continued fixing merge stuff from the earlier branch.

* Fixed some front-end items with Fidu

- Fixed fetching the balance.
- Also fixed the decimal places
- Also fixed when we go fromAtomic in a few places with the balance.

* Refactored all the "fromAtomic" and "toAtomic" functions to be prefixed with usdc (just for clarity)

- Now that we have two ERC20's that we go from / to atomic with, I wanted to clarify which is which,
otherwise I think it could get confusing.

* More refactoring of fromAtomic, and also fixing of units

* A couple small things from self review.

* No longer require approval for burning Fidu.

- This is not necessary when this function is locked down to only MINTER_ROLE
holders. And it greatly hurts the UX, so removing it.

* Fixed some naming again, to be consistent about usage between "decimal places" and "decimals".

- "Decimal places" should literally be the number itself (ie. 6, or 18)
- Decimals should be the large version (ie. 1e6 or 1e18)

* testing if netlify will build

* Fix a few issues from Sanjay's review.

* Removed the es-lint disables.

* Bring the deployPool function back into the main baseDeploy file.

- The original reasoning for extracting it was because we didn't have access
control, and some other refactors weren't in place. Now they are, and no
need to extract it.


* Fix a few methods to better fit the CEI pattern. (#74)

- Found these from an earlier crytic run. Fixing them up
as part of our security enhancements.


* Fix all solhint and compiler warnings

* Additional fixes

* [fix] Additional security fixes (#75)

* Additional security fixes

* Automatically get test contracts by convention

* Add missing initializer

* Improved clarity of how we calculate interest. (#77)

- Turns out there wasn't really a bug with how we calculate interest. But more of
inconsistency with how we were inputting interest rates.
- The inconsistency was causing bugs that were hard to reason about.
- We clarified, and now have a consisten function that takes a number as a string
and as a percentage, and outputs the BN that we should pass to the smart contract.
- Along the way, we found a legitimate bug that would have charged borrowers
interest on their first drawdown that they assuredly should not have been charged,
due to how we were setting the lastUpdatedBlock.
- Specifically, the borrower would have been charged interest for the time in between
having a credit line setup and whenever they did their first drawdown. Woof!
That would not have been a good first experience with the protocol.


* [Feature] - Give us a solid check so minting/burning is always in sync with pool accounting (#78)

* Give us a solid check that minting is never out of whack with the pool accounting.

- Added in a rock solid invariant check that the pool's assets and liabilities always match up
including the proposed mint (under the assumption that the pool should already have received the funds before minting)
- In the process did quite a bit of refactoring to use interfaces, and a "ConfigHelper",
which cleans up a lot of boilerplate stuff we were doing to use the GoldfinchConfig.

* Add another spec for the totalLoansOutstanding number.

* Add a check on burning Fidu as well.

* Fix up a few things from review.


* [Chore] - Added a withdraw script, and cleaned up the deposit script. (#80)

* Added a withdraw script, and cleaned up the deposit script.

- Committing them just in case we need them later.

* Remove unused global ethers.

+++ contracts/test/TestAccountant.sol
@@ -6,8 +6,12 @@ import "../Accountant.sol";
 import "../CreditLine.sol";
 
 contract TestAccountant {
-    function calculateInterestAndPrincipalAccrued(address creditLineAddress, uint blockNumber) public view returns (uint, uint) {
-      CreditLine cl = CreditLine(creditLineAddress);
-      return Accountant.calculateInterestAndPrincipalAccrued(cl, blockNumber);
-    }
+  function calculateInterestAndPrincipalAccrued(address creditLineAddress, uint256 blockNumber)
+    public
+    view
+    returns (uint256, uint256)
+  {
+    CreditLine cl = CreditLine(creditLineAddress);
+    return Accountant.calculateInterestAndPrincipalAccrued(cl, blockNumber);
+  }
 }
